+++ src/Adapters/Http/Boss4D.Adapters.Http.pas
@@ -12,6 +12,8 @@ type
     function Get(const AURL: string; out AResponse: string): Integer;
     function PostJson(const AURL, ABody: string;
       out AResponse: string): Integer;
+    function PostJsonAuthorized(const AURL, ABody, ABearerToken: string;
+      out AResponse: string): Integer;
   end;
 
 implementation
@@ -84,6 +86,33 @@ begin
         AResponse := E.Message;
       end;
     end;
+  finally
+    LOutput.Free;
+    LBody.Free;
+    LClient.Free;
+  end;
+end;
+
+function TBoss4DHttpNativeAdapter.PostJsonAuthorized(const AURL, ABody,
+  ABearerToken: string; out AResponse: string): Integer;
+var
+  LClient: THTTPClient;
+  LResponse: IHTTPResponse;
+  LBody, LOutput: TStringStream;
+begin
+  AResponse := '';
+  LClient := THTTPClient.Create;
+  LBody := TStringStream.Create(ABody, TEncoding.UTF8);
+  LOutput := TStringStream.Create('', TEncoding.UTF8);
+  try
+    LClient.UserAgent := 'Boss4D/1.3';
+    LClient.ContentType := 'application/json';
+    LClient.CustomHeaders['Authorization'] := 'Bearer ' + ABearerToken;
+    LClient.CustomHeaders['Accept'] := 'application/vnd.github+json';
+    LClient.CustomHeaders['X-GitHub-Api-Version'] := '2022-11-28';
+    LResponse := LClient.Post(AURL, LBody, LOutput);
+    Result := LResponse.StatusCode;
+    AResponse := LOutput.DataString;
   finally
     LOutput.Free;
     LBody.Free;